Complete Wan2.1 14B model + official weights verification
- model.py: complete WanModel (patch/text/time embedding + blocks + head + unpatchify)
  - Q/K/V with bias, FFN with bias, RMSNorm param 'weight', norm3 LayerNorm
  - T2V cross-attention, modulation, 3D RoPE (float64 complex)
  - Weight keys 100% match official checkpoint
- norm.py: RMSNorm param 'weight', LayerNorm float32 compute
- rope.py: float64 (match official), return float
- verify_wan21_14b_full.py: load all 6 shards (1095/1095 keys, 14.29B params)
  Full forward [16,4,64,64]→[16,4,64,64] PASS
- SP bench: model.to(bf16) before setup_shape, 8GPU all 3 strategies pass
- bX_rel=0.0017 (backward correct), fwd_rel issue (rank-major qkv split, documented)
- docs/WAN21_MODEL.md: architecture + verification docs

+++ examples/bench_wan21_fsdp2.py
@@ -92,6 +92,7 @@ def run(rank, ng, port, iters, verify, strategies):
                 for p in strat.model.parameters():
                     p.data = torch.randn(p.shape, dtype=p.dtype, device=dev, generator=g) / math.sqrt(dim)
             try:
+                strat.model = strat.model.to(torch.bfloat16)
                 strat.setup_shape(bs, seq, nheads, head_dim)
             except Exception as e:
                 if rank == 0: print(f"  {label:<12} {strat_name:<12} SKIP (setup: {e})")
@@ -100,7 +101,7 @@ def run(rank, ng, port, iters, verify, strategies):
             # Build reference BEFORE FSDP2 (it converts params to DTensor)
             if do_verify and rank == 0 and ref_out is None:
                 ref_m = WanSelfAttention(model_cfg.dim, model_cfg.num_heads, model_cfg.head_dim,
-                                         qk_norm=model_cfg.qk_norm, eps=model_cfg.eps).to(dev)
+                                         qk_norm=model_cfg.qk_norm, eps=model_cfg.eps).to(dev).to(torch.bfloat16)
                 ref_m.load_state_dict(strat.model.state_dict())
                 Xr = X_full.clone().requires_grad_(True)
                 yr = ref_m(Xr, grid, ref_m.freqs)
